backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
duckdb==1.1.3
pandas==2.2.3
requests==2.32.3
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
numpy==2.1.3
